finally handling user input correctly

+++ MathStuff.java
@@ -1,10 +1,10 @@
 public class MathStuff{
-  long x;
-  long y;
-  long temp;
+  int x;
+  int y;
+  int temp;
   //TYPE 1 = X, TYPE 2 = Y
 
-  public void setValue(long valArg, int type){
+  public void setValue(int valArg, int type){
     switch(type){
       case 1:
         x = valArg;
@@ -16,7 +16,7 @@ public class MathStuff{
     }
   }
 
-  public void addNumber(long addNum, int type){
+  public void addNumber(int addNum, int type){
     switch(type){
       case 1:
         x = x + addNum;
@@ -35,7 +35,7 @@ public class MathStuff{
     }
   }
 
-  public void thisFunctionThatIUseALot(long valArg, long addNum, int type){ //I use this function a lot
+  public void thisFunctionThatIUseALot(int valArg, int addNum, int type){ //I use this function a lot
     setValue(valArg, type);
     addNumber(addNum, type);
     getValue(type);
